Fix pointer mistake which caused segfaults
