+++ srcs/lexer/lexer_tokenlstiter.c
@@ -1,0 +1,21 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        ::::::::            */
+/*   lexer_tokenlstiter.c                               :+:    :+:            */
+/*                                                     +:+                    */
+/*   By: jbrinksm <jbrinksm@student.codam.nl>         +#+                     */
+/*                                                   +#+                      */
+/*   Created: 2019/10/31 19:58:31 by jbrinksm       #+#    #+#                */
+/*   Updated: 2019/10/31 19:58:42 by jbrinksm      ########   odam.nl         */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "vsh.h"
+
+void		lexer_tokenlstiter(t_tokenlst *lst, void (*f)(t_tokenlst *elem))
+{
+	if (lst == NULL || f == NULL)
+		return ;
+	f(lst);
+	lexer_tokenlstiter(lst->next, f);
+}
